refactor(website): implement DDD architecture and optimize SSG
- Move components to proper feature modules (DDD structure)
- Update generateStaticParams to use routing configuration
- Maintain all tests passing (87/87)
