+++ apps/website/app/[locale]/page.tsx
@@ -1,15 +1,16 @@
-import { HomePage } from "@/components/home/home-page";
+import { HomePage } from "@/lib/features/home/components/home-page";
 
 interface PageProps {
   params: Promise<{ locale: string }>;
 }
 
-// SSG: 預生成所有支援的語言路由
+import { routing } from '@/i18n/routing';
+
+// SSG: 從 i18n 配置自動生成支援的語言路由
 export async function generateStaticParams() {
-  return [
-    { locale: 'en' },
-    { locale: 'zh' }
-  ];
+  return routing.locales.map((locale) => ({
+    locale
+  }));
 }
 
 export default async function Page({ params }: PageProps) {
