+++ apps/website/lib/features/about/components/about-hero.tsx
