+++ apps/website/lib/features/about/components/core-principles.tsx
