+++ apps/website/lib/features/about/components/core-values.tsx
