+++ apps/website/lib/features/home/components/about-section.tsx
